N6700: Check stat:oper:cond for trigger initiated
Previously the Status Operation Event register (stat:oper) was
checked to know when the instrument is ready to receive a
trigger after initiating. This is wrong since the event
register is clear-on-read.
Instead we should use the Status Operation Condition
register (stat:oper:cond) which holds the live
(unlatched) operational status of the instrument.

This would manifest in data not being available when we expect it
to be since the trigger system was not initiated properly
before sending the trigger.